Feat : OW-66 파일 수정 API 및 테스트 추가
- 파일 삭제 기능 구현
  - Controller 테스트 추가
  - adoc 갱신 및 정리
- 코드 정리, 수정
  - FileController 메서드명 수정
  - S3File/FileService에 누락된 @Transactional 추가
  - file.controller 패키지 생성, FileControllerTest 패키지 이동
  - FileService에서 컨테이너가 존재하는지 확인하는 로직 메서드 추출

+++ back/src/main/java/com/ogjg/back/file/service/FileService.java
@@ -4,6 +4,7 @@ import com.ogjg.back.container.exception.NotFoundContainer;
 import com.ogjg.back.container.repository.ContainerRepository;
 import com.ogjg.back.file.dto.request.CreateFileRequest;
 import com.ogjg.back.file.dto.request.DeleteFileRequest;
+import com.ogjg.back.file.dto.request.UpdateFileRequest;
 import com.ogjg.back.s3.service.S3FileService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -20,17 +21,29 @@ public class FileService {
     @Transactional
     public void createFile(String loginEmail, CreateFileRequest request) {
         String filePath = request.getFilePath();
-        containerRepository.findByNameAndEmail(extractContainerName(filePath), loginEmail)
-                .orElseThrow(NotFoundContainer::new);
+        if (!isContainerExist(loginEmail, extractContainerName(filePath))) throw new NotFoundContainer();
 
         s3FileService.createFile(loginEmail, filePath);
     }
 
+    @Transactional
     public void deleteFile(String loginEmail, DeleteFileRequest request) {
         String filePath = request.getFilePath();
-        containerRepository.findByNameAndEmail(extractContainerName(filePath), loginEmail)
-                .orElseThrow(NotFoundContainer::new);
+        if (!isContainerExist(loginEmail, extractContainerName(filePath))) throw new NotFoundContainer();
 
         s3FileService.deleteFile(loginEmail, filePath);
+    }
+
+    @Transactional
+    public void updateFile(String loginEmail, UpdateFileRequest request) {
+        String filePath = request.getFilePath();
+        if (!isContainerExist(loginEmail, extractContainerName(filePath))) throw new NotFoundContainer();
+
+        s3FileService.updateFile(loginEmail, filePath, request.getContent());
+    }
+
+    private boolean isContainerExist(String loginEmail, String containerName) {
+        return containerRepository.findByNameAndEmail(containerName, loginEmail)
+                .isPresent();
     }
 }
